+++ docs/context/interface/api.md
@@ -26,6 +26,9 @@ sources:
   - src/treg/routers/web.py
   - src/treg/domain/identity/access.py
   - src/treg/domain/governance/teams.py
+  - src/treg/domain/governance/access.py
+  - src/treg/domain/governance/budgets.py
+  - src/treg/domain/governance/publicdemo.py
   - src/treg/domain/identity/mcp_oauth.py
   - src/treg/domain/identity/session.py
   - src/treg/timeutil.py
@@ -484,7 +487,8 @@ validated before resolving the shared HTTP client. `/auth/logout` remains an HTT
   (every mutation is frozen no matter what routes are added later), and `require_identity` refuses the
   token entirely (it must never act as a user — mint identity tokens, create orgs, accept invites). Its
   `/call` traffic is metered per client IP (`_enforce_public_demo_ip_cap`, `PUBLIC_DEMO_HIT_NS`,
-  ~10 calls/min/IP) since one token stands in for thousands of strangers.
+  ~10 calls/min/IP) since one token stands in for thousands of strangers. The limiter and its
+  constants share the `domain.governance.publicdemo` owner.
 - **Skills / bundles:** `register_skill` (`POST /skills`) composes a `Bundle` + its secrets + tools
   atomically, resolving each binding's `secret` local-name to the created secret id; the shared core is
   `_register_skill_bundle` (also used by the folder importer). `list_bundles`, `get_bundle`,
